code_memory_system/pyproject.toml
[tool.poetry]
name = "code-memory-system"
version = "0.1.0"
description = "Intelligent code memory system for AI-assisted development"
authors = ["AI Assistant"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0"
chromadb = "^0.4.0"
click = "^8.1"
pydantic = "^2.0"
tree-sitter = "^0.20"
openai = "^1.0"
redis = {version = "^5.0", optional = true}
aiofiles = "^23.0"
tqdm = "^4.66"
python-dotenv = "^1.0"
numpy = "^1.26"
scikit-learn = "^1.3"
watchdog = "^3.0"
pyyaml = "^6.0"
pytest = "^7.4"
black = "^23.0"
mypy = "^1.7"
pytest-asyncio = "^0.21"

[tool.poetry.extras]
redis = ["redis"]

[tool.poetry.scripts]
code-memory = "src.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true